+++ stackslib/src/clarity_vm/clarity.rs
@@ -215,7 +215,7 @@ impl<'a, 'b> ClarityBlockConnection<'a, 'b> {
     pub fn cost_so_far(&self) -> ExecutionCost {
         match self.cost_track {
             Some(ref track) => track.get_total(),
-            None => ExecutionCost::zero(),
+            None => ExecutionCost::ZERO,
         }
     }
 
@@ -1765,7 +1765,7 @@ impl<'a, 'b> ClarityTransactionConnection<'a, 'b> {
     pub fn cost_so_far(&self) -> ExecutionCost {
         match self.cost_track {
             Some(ref track) => track.get_total(),
-            None => ExecutionCost::zero(),
+            None => ExecutionCost::ZERO,
         }
     }