+++ stacks-signer/CHANGELOG.md
@@ -7,7 +7,25 @@ and this project adheres to the versioning scheme outlined in the [README.md](RE
 
 ## [Unreleased]
 
+## Added
+
+## Changed
+- Signers now listen to new block events from the stacks node to determine whether a block has been successfully appended to the chain tip
+
+## [3.1.0.0.1.0]
+
 ### Added
+
+### Changed
+
+- Added tenure extend timestamp to signer block responses
+- Added tenure_idle_timeout_secs configuration option for determining when a time-based tenure extend will be accepted
+
+## [3.1.0.0.0.0]
+
+### Added
+
+- **SIP-029 consensus rules, activating in epoch 3.1 at block 875,000** (see [SIP-029](https://github.com/will-corcoran/sips/blob/feat/sip-029-halving-alignment/sips/sip-029/sip-029-halving-alignment.md) for details)
 
 ### Changed